minor fix for small phone, cleanup
