--- berkeley-mobile/StudyPact/StudyGroups/StudyGroupMember.swift
@@ -1,17 +1,0 @@
-//
-//  StudyGroupMember.swift
-//  berkeley-mobile
-//
-//  Created by Shawn Huang on 1/21/21.
-//  Copyright © 2021 ASUC OCTO. All rights reserved.
-//
-
-import UIKit
-
-struct StudyGroupMember {
-    let profilePictureURL: URL?
-    var profilePicture: UIImage?
-    let name: String
-    let email: String
-    let phoneNumber: String?
-}
